% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_chat_sheet.r
\name{open_chat_sheet}
\alias{open_chat_sheet}
\title{Create a new Chat Sheet}
\usage{
open_chat_sheet(path = "chat.Rmd")
}
\arguments{
\item{path}{\code{character(1)} Destination path for the new R Markdown file. Use
\code{NULL} (default) to let the function choose an available filename in the
working directory.}
}
\value{
Invisibly returns the path to the file that was created.
}
\description{
Open a new chat sheet file.
}
