% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-sheet.R
\name{parse_chat_sheet}
\alias{parse_chat_sheet}
\title{Parse a prompt sheet and returns a dataframe with columns
\itemize{
\item role: one of 'system' 'user' or assistant'
\item turn: integer, the turn number in the sequence. System prompt if present is always 0.
user turns are always odd, llm turns are even.
\item text:
}}
\usage{
parse_chat_sheet(file, text = readLines(file))
}
\description{
Parse a prompt sheet and returns a dataframe with columns
\itemize{
\item role: one of 'system' 'user' or assistant'
\item turn: integer, the turn number in the sequence. System prompt if present is always 0.
user turns are always odd, llm turns are even.
\item text:
}
}
