% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_chat_sheet}
\alias{run_chat_sheet}
\title{Run a chat prompt sheet}
\usage{
run_chat_sheet(path = NULL, stream = TRUE)
}
\arguments{
\item{path}{Path to the \code{chat.Rmd} file. If \code{NULL}, uses the current active
document in RStudio.}

\item{stream}{Whether to stream the response as it's generated (\code{TRUE}) or
return the full response at once (\code{FALSE}).}
}
\value{
Invisibly returns the updated \code{ellmer::Chat} object.
}
\description{
This calls \code{read_chat_sheet()}, uses the generated \code{ellmer::Chat} object to
submit a request to the LLM provider, and streams the response back into
originating document.
}
